# packages/contracts/openapi/api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnePic API",
    "version": "0.1.0",
    "summary": "OnePic 单图模板库的首期受控生图 API",
    "description": "当前 phase-one API 表面（docs/full-stack-architecture.md §6、ADR 0001 D-1）：\n健康、上传、预审、生成/取消/删除/历史、收藏/集合与工作区导出。模板目录仍由\n静态 JSON/TXT/WebP 提供。成功响应为 `{ data, meta }`，失败响应为 `{ error }`；\n字段命名 camelCase，时间 UTC ISO-8601，ID 使用 UUID/ULID，分页使用 cursor。\n本契约是生成类型与运行时校验的唯一事实来源；实现和 generated client 漂移由 CI 阻断。\n"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/auth/callback": {
      "get": {
        "operationId": "completeLogin",
        "tags": [
          "auth"
        ],
        "summary": "完成 OIDC 授权码回调",
        "description": "校验 code、state、onepic_login 挑战与 OIDC token 后创建 opaque 会话； 成功时设置 onepic_session、清除 onepic_login，并重定向到站点根路径。\n",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/AuthCode"
          },
          {
            "$ref": "#/components/parameters/AuthState"
          },
          {
            "$ref": "#/components/parameters/PendingLoginCookie"
          }
        ],
        "responses": {
          "302": {
            "description": "登录成功并重定向到站点根路径",
            "headers": {
              "Location": {
                "description": "登录后的站内重定向路径。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "getCurrentSubject",
        "tags": [
          "auth"
        ],
        "summary": "查询当前登录主体",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "当前登录主体",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfAuthMe"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refreshSession",
        "tags": [
          "auth"
        ],
        "summary": "轮换当前会话",
        "description": "签发新 opaque session cookie 并撤销旧会话。",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "会话已轮换",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfAuthRefresh"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "tags": [
          "auth"
        ],
        "summary": "注销当前会话",
        "description": "有会话时撤销它，并始终清除 onepic_session cookie。",
        "security": [
          {
            "sessionCookie": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "注销完成",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfAuthLogout"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "onepic_session",
        "description": "服务端 opaque 会话（OIDC 授权码 + PKCE 登录后签发，B03）。"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "x-onepic-requested-with",
        "in": "header",
        "required": true,
        "description": "CSRF 护栏自定义头，恒为 onepic-fetch（跨站表单无法设置）。",
        "schema": {
          "type": "string",
          "const": "onepic-fetch"
        }
      },
      "AuthCode": {
        "name": "code",
        "in": "query",
        "required": true,
        "description": "OIDC provider 返回的一次性授权码。",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "AuthState": {
        "name": "state",
        "in": "query",
        "required": true,
        "description": "与服务端 pending login challenge 绑定的 OIDC state。",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "PendingLoginCookie": {
        "name": "onepic_login",
        "in": "cookie",
        "required": true,
        "description": "/auth/login 设置的短期、HttpOnly 登录挑战句柄。",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
...
    },
    "responses": {
...
      "Unauthenticated": {
        "description": "无有效会话",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      },
      "ApiFailureResponse": {
        "description": "稳定错误码信封（错误码目录见 D06 数据字典）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiMeta": {
        "type": "object",
        "description": "与 data 并列的分页/追踪元数据；结构随端点补充。",
        "additionalProperties": true
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "稳定错误码信封；客户端不得依赖 message 文本做逻辑判断。",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "correlationId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "稳定机器可读错误码（错误码目录见 D06 数据字典）。"
          },
          "message": {
            "type": "string",
            "description": "面向人类的错误说明；不得包含密钥或提示词正文。"
          },
          "details": {
            "type": "object",
            "description": "结构化错误细节；不得包含密钥。",
            "additionalProperties": true
          },
          "correlationId": {
            "type": "string",
            "description": "请求关联 ID，用于日志与支持排查。"
          }
        }
      },
      "ApiFailure": {
        "type": "object",
        "description": "失败响应信封。",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        }
      },
      "Subject": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "issuer",
          "subjectClaim",
          "role"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "issuer": {
            "type": "string",
            "format": "uri"
          },
          "subjectClaim": {
            "type": "string",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "enum": [
              "guest",
              "member",
              "admin"
            ]
          }
        }
      },
      "AuthMe": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "subject"
        ],
        "properties": {
          "subject": {
            "$ref": "#/components/schemas/Subject"
          }
        }
      },
      "AuthRefresh": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "rotated"
        ],
        "properties": {
          "rotated": {
            "type": "boolean",
            "const": true
          }
        }
      },
      "AuthLogout": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "loggedOut"
        ],
        "properties": {
          "loggedOut": {
            "type": "boolean",
            "const": true
          }
        }
      },
      "ApiSuccessOfAuthMe": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthMe"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiSuccessOfAuthRefresh": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthRefresh"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiSuccessOfAuthLogout": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthLogout"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
    }
  }
}